+++ persistent/ChangeLog.md
@@ -46,6 +46,13 @@
     constraints on an entity update without having to update it.
 * [#1142](https://github.com/yesodweb/persistent/pull/1142)
     * Deprecate `hasCompositeKey` in favor of `hasCustomPrimaryKey` and `hasCompositePrimaryKey` functions.
+* [#1098](https://github.com/yesodweb/persistent/pull/1098)
+  * Add support for configuring the number of stripes and idle timeout for connection pools 
+    * For functions that do not specify an idle timeout, the default has been bumped to 600 seconds.
+      * This change is based off the experience of two production codebases. See [#775](https://github.com/yesodweb/persistent/issues/775)
+    * Add a new type `ConnectionPoolConfig` to configure the number of connections in a pool, their idle timeout, and stripe size.
+    * Add `defaultConnectionPoolConfig` to create a `ConnectionPoolConfig`
+    * Add `createSqlPoolWithConfig` and `withSqlPoolWithConfig`, which take this new data type
 
 ## 2.10.5.2